dev: 0.3.1: adding appointment, list+filter appointments, appointments statuses

+++ database/migrations/information/branches/2023_11_22_213045_create_branch_services_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('branch_id')->references('id')->on('branches')->onDelete('cascade');
             $table->foreignId('service_id')->references('id')->on('services')->onDelete('cascade');
-            $table->decimal('price', 10, 2);
+            $table->bigInteger('price');
             $table->timestamps();
 
             $table->unique(['branch_id', 'service_id']);
